+++ R/plot_heatmap.R
@@ -10,13 +10,13 @@ setMethod(
   signature = signature(object = "matrix"),
   definition = function(object, col = grDevices::hcl.colors(12, "YlOrBr", rev = TRUE),
                         diag = TRUE, upper = TRUE, lower = TRUE,
-                        freq = FALSE, margin = 1,
+                        freq = FALSE, margin = 1, fixed_ratio = TRUE,
                         axes = TRUE, legend = TRUE, ...) {
 
     plot_matrix(object, panel = panel_tiles, col = col,
                  diag = diag, upper = upper, lower = lower,
                  freq = freq, margin = margin, drop_zero = FALSE,
-                 axes = axes, legend = legend)
+                 axes = axes, legend = legend, asp = fixed_ratio)
 
     invisible(object)
   }
@@ -30,12 +30,13 @@ setMethod(
   signature = signature(object = "data.frame"),
   definition = function(object, col = grDevices::hcl.colors(12, "YlOrBr", rev = TRUE),
                         diag = TRUE, upper = TRUE, lower = TRUE,
-                        freq = FALSE, margin = 1,
+                        freq = FALSE, margin = 1, fixed_ratio = TRUE,
                         axes = TRUE, legend = TRUE, ...) {
     object <- data.matrix(object)
     methods::callGeneric(object, col = col,
                          diag = diag, upper = upper, lower = lower,
                          freq = freq, margin = margin,
+                         fixed_ratio = fixed_ratio,
                          axes = axes, legend = legend)
   }
 )